Move resizing and rotating to separate files, too.
